+++ NEWS.md
@@ -16,6 +16,9 @@ which always resolves to the latest release.
     - Test coverage remains at 100%.
 
 
+* Paths to Exiobase energy flows now added.
+
+
 # PFUSetup 0.1.14 (2023-08-18) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.8264422.svg)](https://doi.org/10.5281/zenodo.8264422)
 
 * Now uses DOI for long-term citation.